Starting to get an experiment of TX events...
Not great but every 250ms I try to output a event saying that one of the storage info changed...

I am starting to see when I copy file in, that if go look at the top level root area, the remaining size is updated... Not sure yet if this is doing it or not.

I am also not always getting any  tx callback.  so maybe need to wait until previous one...