MTP process some messages early Debug Stream
The mtpd.begin() now starts up interval timer that will go away the first time you call loop,  That allows it to process a few host messages quickly to hopefully not have the PC time out MTP while it is starting up.

With this also created quck and dirty memory stream, that can capture stuff before the USB Serial is ready.   Also added method to MTPD to set which stream output should go to and updated storage and some others to look at this.

Still WIP, some of this will likely be removed.