Fix MTP date issue when creating a new directory on PC
There were issues with both files and folders.  But with Files, we alwarys overwrote it with the date that came from MTP... SO did not see it.

With folders we did not retrieve the dates from the newly created folder and as such when host asked for it we returned 0 or what that actually converts to